refactor: 💡 update datepicker component, add form validation
