+++ src/common.module.scss
@@ -1,3 +1,5 @@
+$twitterBlue: #1d9bf0;
+$twitterGray: #86818c;
 .input {
     border: solid 1px #cfd9de;
     height: 53px;
