+++ src/containers/Register/index.module.scss
@@ -1,5 +1,4 @@
-@value twitterBlue #1d9bf0;
-@value twitterGray #86818c;
+@import '../../common.module.scss';
 
 .form {
     margin: 0 30px;
@@ -10,8 +9,14 @@
         background: none;
         border: none;
     }
+    :global(.adm-list-item) {
+        padding: 0;
+        margin: 0;
+    }
     :global(.adm-list-item-content) {
         border: none;
+        padding: 0;
+        margin: 0;
     }
 }
 
@@ -23,11 +28,11 @@
 }
 
 .input {
-    composes: input className from '../../common.module.css';
+    composes: input className from '../../common.module.scss';
 }
 
 .changeTypeButton {
-    color: twitterBlue;
+    color: $twitterBlue;
 }
 
 .birthdayTitle{
@@ -47,7 +52,7 @@
 .footerButton {
     margin: 0 auto;
     border-radius: 50px;
-    background-color: twitterGray;
+    background-color: $twitterGray;
     color: #fff;
     font-size: 16px;
     width: 80%;
