+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 // import Login from './containers/Login';
 import Register from './containers/Register';
-import { startVconsole } from './utils';
+// import { startVconsole } from './utils';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
@@ -13,4 +13,4 @@ root.render(
 );
 
 // 启动 vconsole
-startVconsole();
+// startVconsole();
